Initial Commit - Persists an http request that has an email into a AWS DynamoDB record
General Overview:

This is a pure rust package that utilizes the experimental rust lambda
runtime to mediate between aws and application code. In this case, the
application receives an email in a post request, validates the email,
validates that the email doesn't already exist in the dynamodb table,
and then commits it to the table. This provides a very raw "email" list
that lends itself to ease of use in the future. The email record is
accompanied by a "DateReceived" number in Epoch seconds that captures
when the service received this email.  This can be useful for any
DSAR/IDD as emails are generally considered Personally identifiable
information (PII).

Problem:

Voto is an ambitious project with lots of potential feature areas.
Currently in its infancy, there is no way for anyone to express interest
or communicate feature requests.

Solution:

This commit implements a lambda that stores interested users' email in a
dynamodb table. This is a crude implementation that could likely be
implemented with less effort in a different manner, but implementing
this in the aws ecosystem positions the project for ease of
extensibility later.